requirements.txt
streamlit>=1.32.0
tweepy>=4.14.0
requests>=2.32.0
streamlit-autorefresh>=1.0.1
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
